scale_error moved to scale library
