+++ core/primitives/CMakeLists.txt
@@ -8,11 +8,10 @@ add_library(primitives
         scale_codec.hpp
         impl/scale_codec_impl.hpp
         impl/scale_codec_impl.cpp
-        scale_error.hpp  # these 2 files will be moved
-        scale_error.cpp  # to scale library
         )
 
 target_link_libraries(primitives
         buffer
         outcome
+        scale
         )
