+++ core/primitives/impl/scale_codec_impl.cpp
@@ -9,8 +9,7 @@
 #include "scale/collection.hpp"
 #include "scale/compact.hpp"
 #include "scale/fixedwidth.hpp"
-// TODO(yuraz): change path after moving scale_error to scale
-#include "primitives/scale_error.hpp"
+#include "scale/scale_error.hpp"
 
 using namespace kagome::common::scale;
