+++ core/scale/scale_error.hpp
